+++ CHANGELOG.md
@@ -13,6 +13,8 @@
   value of `Compression`, now set by default to `Compression::None` (no compression) and which can always be changed
   with the `compression()` and (indirectly) `filename()` methods. For this reason, the `BackupExport::$extension`
   property no longer exists;
+* the `BackupImport::filename()` method uses `Compression` to check the validity of the file you want to import (so it
+  no longer checks its extension). This will throw a `ValueError` exception for invalid files;
 * the `RotateCommand` class is deprecated and will be removed in a later release. For this reason, the `ExportCommand`
   class now uses the `BackupManager::rotate()` method to continue supporting the `--rotate` option;
 * the `BackupTrait::getValidCompressions()` method is deprecated. Will be removed in a future release;